spawn strategic and bases classes
this breakes up the spawning on strategic and airbases into two different
classes and then orders them such that the template the airbase attempts
to spawn have already been read.

This doesn't make sense though because the generate function is executed
well after all templates are read so there is no need to order airbases
after other strategic spawning.

Finally, the concept of spawning an airbase and all of its subordinate
units is not a good one when considering state saving. It instead would
be better to provide a special method that airbase objects provide to
generate all their subordinate assets. This way they are only called
at generation time. We currently do not support bases being captured,
just killed for now.